+++ src/modules/developer-example/module.js
@@ -17,7 +17,7 @@ const developerExampleModule = {
     },
   },
   category: "developer",
-  version: "0.32.3.2",
+  version: "0.32.3.3",
   enabledByDefault: false,
   canDisable: true,
   historicalReadAccess: true,